Rename files to SpubHeader naming convention
